Break test code in multiple files
Jira: https://issues.redhat.com/browse/ASC-441
